requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.18.0+cpu
torch==2.3.0+cpu
gunicorn
Werkzeug
Flask
requests
Pillow